feature: pr updates and storybook edits
