--- .babelrc
@@ -1,14 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-react"
-  ],
-  "plugins": [
-    [
-      "@babel/plugin-transform-react-jsx",
-      {
-          "runtime": "automatic"
-      }
-    ]
-  ]
-}
