+++ .storybook/main.ts
@@ -10,7 +10,7 @@ module.exports = {
   },
   stories: [
     '../src/**/*.stories.mdx',
-    '../src/**/*.stories.@(js|jsx|ts|tsx)'
+    '../**/*.stories.@(js|jsx|ts|tsx)'
   ],
   addons: [
     '@storybook/addon-links',
