--- src/Example/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Example.js'
