+++ src/components/ConsentToggle/ConsentToggle.stories.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import ConsentToggle from "./ConsentToggle.js";
+import type { ConsentToggleProps } from './ConsentToggle.js';
 
 export default {
     title: "Metrics Consent Toggle",
@@ -7,7 +8,7 @@ export default {
     args: {}
 };
 
-const Template = args => <ConsentToggle {...args}/>;
+const Template = (args:ConsentToggleProps) => <ConsentToggle {...args}/>;
 
 const Standard = Template.bind({});
