--- src/types/global.d.ts
@@ -1,3 +1,0 @@
-interface Window {
-  Countly: unknown
-}
