+++ tsconfig.json
@@ -1,24 +1,10 @@
 {
   "extends": "aegir/src/config/tsconfig.aegir.json",
   "compilerOptions": {
-    "allowSyntheticDefaultImports": true,
-    "declaration": true,
-    "esModuleInterop": true,
-    // "skipLibCheck": true,
     "jsx": "react",
-    "lib": ["ES5", "ES2015", "ES2016", "DOM", "ESNext"],
     "types": ["node"],
-    // "module": "ESNext",
-    // "moduleResolution": "node",
-    "noImplicitAny": false,
-    "noUnusedLocals": true,
     "outDir": "dist",
-    "sourceMap": true,
-    "strict": true,
     "target": "ES6",
-    // "declarationDir": "types",
-    // "emitDeclarationOnly": true,
-    "typeRoots": ["node_modules/@types", "src/types"],
   },
   "include": [
     "src",
@@ -26,10 +12,6 @@
   ],
   "exclude": [
     "node_modules",
-    "dist",
-    "node_modules",
-    "src/**/*.test.tsx",
-    "src/**/*.stories.tsx",
   ],
   "ts-node": {
     "extends": "aegir/src/config/tsconfig.aegir.json",
